+++ src/main/kotlin/no/nav/familie/pdf/pdf/PdfElementUtils.kt
@@ -3,11 +3,16 @@ package no.nav.familie.pdf.pdf
 import com.itextpdf.io.image.ImageDataFactory
 import com.itextpdf.kernel.colors.DeviceRgb
 import com.itextpdf.kernel.pdf.tagging.StandardRoles
+import com.itextpdf.layout.borders.Border
+import com.itextpdf.layout.element.Cell
+import com.itextpdf.layout.element.Div
 import com.itextpdf.layout.element.Image
 import com.itextpdf.layout.element.List
 import com.itextpdf.layout.element.ListItem
 import com.itextpdf.layout.element.Paragraph
+import com.itextpdf.layout.element.Table
 import com.itextpdf.layout.element.Text
+import com.itextpdf.layout.properties.UnitValue
 import no.nav.familie.pdf.pdf.domain.VerdilisteElement
 
 object PdfElementUtils {
@@ -69,7 +74,7 @@ object PdfElementUtils {
             accessibilityProperties.role = StandardRoles.P
         }
 
-    fun sjekkDobbelLinjeskift(tekst: String): String {
+    private fun sjekkDobbelLinjeskift(tekst: String): String {
         val bareLinjeskiftRegex = Regex("^\\n+$")
         val dobbelLinjeskiftRegex = Regex("\\n{2,}")
         val rensetVerdi =
@@ -104,5 +109,79 @@ object PdfElementUtils {
             setFontSize(tekstStørrelse)
             simulateBold()
             accessibilityProperties.role = rolle
+        }
+
+    fun lagTabell(
+        tabellData: VerdilisteElement,
+    ): Table {
+        requireNotNull(tabellData.verdiliste) { "VerdilisteElement må ha en verdiliste for å kunne lage en tabell" }
+
+        val tabell =
+            Table(UnitValue.createPercentArray(floatArrayOf(50f, 50f))).apply {
+                useAllAvailableWidth()
+                setMarginBottom(10f)
+                setMarginLeft(15f)
+                accessibilityProperties.role = StandardRoles.TABLE
+            }
+
+        val captionDiv =
+            Div().apply {
+                add(
+                    Paragraph(tabellData.label).apply {
+                        setFontColor(DeviceRgb(0, 52, 125))
+                        setFontSize(14f)
+                        simulateBold()
+                    },
+                )
+            }
+        tabell.caption = captionDiv
+        tabell.addCell(lagTabellOverskriftscelle("Spørsmål"))
+        tabell.addCell(lagTabellOverskriftscelle("Svar", false))
+        tabell.lagTabellRekursivt(tabellData.verdiliste)
+        return tabell
+    }
+
+    private fun lagTabellOverskriftscelle(
+        tekst: String,
+        erVenstreKolonne: Boolean = true,
+    ): Cell =
+        Cell().apply {
+            add(
+                Paragraph(tekst).apply {
+                    setFontColor(DeviceRgb(0, 86, 180))
+                    setFontSize(14f)
+                    simulateBold()
+                },
+            )
+            setBorder(Border.NO_BORDER)
+            if (erVenstreKolonne) setPaddingRight(10f) else setPaddingLeft(10f)
+            accessibilityProperties.role = StandardRoles.TH
+        }
+
+    private fun Table.lagTabellRekursivt(
+        tabellData: kotlin.collections.List<VerdilisteElement>,
+    ) {
+        tabellData.forEach { element ->
+            when {
+                element.verdi != null -> {
+                    this.addCell(lagTabellInformasjonscelle(element.label, erUthevet = true))
+                    this.addCell(lagTabellInformasjonscelle(element.verdi, erVenstreKolonne = false))
+                }
+                element.verdiliste != null -> lagTabellRekursivt(element.verdiliste)
+            }
+        }
+    }
+
+    private fun lagTabellInformasjonscelle(
+        tekst: String,
+        erVenstreKolonne: Boolean = true,
+        erUthevet: Boolean = false,
+    ): Cell =
+        Cell().apply {
+            add(Paragraph(tekst).setFontSize(12f))
+            setBorder(Border.NO_BORDER)
+            if (erUthevet) simulateBold()
+            if (erVenstreKolonne) setPaddingRight(10f) else setPaddingLeft(10f)
+            accessibilityProperties.role = StandardRoles.TD
         }
 }
